Enable multi-DDL transaction blocks
This is a pretty hacky solution, but it's only a couple of lines to get
here after changing the global connection list into the hash in prior
commits. We should be removing this shortly anyhow.

Also updated a test that checked multi-DDL blocks fail; since they no
longer do, test can check that they are allowed, instead.